fix(layout): won't keep scroll state in `keep-alive` component, closes #3054
